backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
aiofiles>=23.2.1
PyPDF2>=3.0.1
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
google-generativeai>=0.3.2
httpx>=0.26.0
